# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Visit Lab",
        "contact": {
            "email": "saviogmoiagaia@gmail.com"
        },
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/pre-registrations": {
            "get": {
                "tags": [
                    "Pre registro"
                ],
                "summary": "Retorna os pre registro feitos pelo usuario",
                "operationId": "f54ebc59cbd07913ac452384f29e4380",
                "responses": {
                    "200": {
                        "description": "Retorna os pre registro feitos pelo usuario"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/pre-registrations/my": {
            "get": {
                "tags": [
                    "Pre registro"
                ],
                "summary": "Retorna o pre registro para o usuario",
                "operationId": "8fdef2f4ae4c193b9f367ea02ca9009e",
                "responses": {
                    "200": {
                        "description": "Retorna o pre registro para o usuario"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/pre-registrations/": {
            "post": {
                "tags": [
                    "Pre registro"
                ],
                "summary": "Cria um pré registro de usuário",
                "description": "Cria um pre registro para proximo login de um usuário",
                "operationId": "a8b315528bb9d3dc581af2374fc3979a",
                "requestBody": {
                    "description": "Json informações necessárias",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "role_id"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Login do usuario",
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "description": "Role que será atribuída a um usuário",
                                        "type": "number"
                                    },
                                    "laboratory_id": {
                                        "description": "Laboratório que será atribuído do usuario",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Login não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/pre-registrations/{preRegistration}": {
            "patch": {
                "tags": [
                    "Pre registro"
                ],
                "summary": "Usuario aceita ou rejeita um cargo",
                "description": "Atualiza o status de um pre registro",
                "operationId": "8e5ff4a0343a98e8c8fc9fcedbb58afd",
                "parameters": [
                    {
                        "name": "preRegistration",
                        "in": "path",
                        "description": "id do preregistro",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Json informações necessárias",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "response"
                                ],
                                "properties": {
                                    "response": {
                                        "description": "Resposta do usuario",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "formulario invalido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Access token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "description": "Chave de aplicação para autenticação na API",
                "name": "X-Application-Key",
                "in": "header"
            }
        }
    }
}